third_party/blink/web_tests/http/tests/devtools/sources/autocomplete-css-expected.txt

The test verifies autocomplete suggestions for CSS file.


Running: testClassNameAutocomplete
========= Selection In Editor =========
.red { color: red }
.blue { color: blue }
.|
======= Autocomplete Suggestions =======
[.red, .blue]

Running: testPropertyNameAutocomplete1
========= Selection In Editor =========
.red { color: red }
.blue { c|
======= Autocomplete Suggestions =======
[color]

Running: testPropertyNameAutocomplete2
========= Selection In Editor =========
.my-class { -|webkit-border: 1px solid black; -webkit-color: blue;
text-align: }
======= Autocomplete Suggestions =======
[-webkit-color]

Running: testPropertyValueAutocomplete1
========= Selection In Editor =========
.red { border-style: |
/* some other words to mess up */
======= Autocomplete Suggestions =======
[dashed, dotted, double, groove, hidden, inherit, initial, inset, none, outset, ridge, solid, unset]

Running: testPropertyValueAutocomplete2
========= Selection In Editor =========
.red { border-style: d|
/* dial drummer dig */
======= Autocomplete Suggestions =======
[dashed, dotted, double]

Running: testPropertyValueAutocomplete3
========= Selection In Editor =========
.red { border-style: z|
/* zipper zorro zion */
======= Autocomplete Suggestions =======
[]

Running: testPropertyValueAutocomplete4
========= Selection In Editor =========
.red { border-style/* comment */: /* comment */|
======= Autocomplete Suggestions =======
[dashed, dotted, double, groove, hidden, inherit, initial, inset, none, outset, ridge, solid, unset]

Running: testPropertyValueAutocomplete5
========= Selection In Editor =========
.my-class { -webkit-border: 1px solid black; -webkit-color: blue;
text-align: |}
======= Autocomplete Suggestions =======
[center, end, inherit, initial, justify, left, right, start, unset, -webkit-auto, -webkit-center, -webkit-left, -webkit-match-parent, -webkit-right]

Running: verifySuggestionsOnColumnTypedIn
.green {
   display|
Suggestions displayed on ':' symbol typed in:
[block, contents, flex, flow-root, grid, inherit, initial, inline, inline-block, inline-flex, inline-grid, inline-table, list-item, none, table, table-caption, table-cell, table-column, table-column-group, table-footer-group, table-header-group, table-row, table-row-group, unset, -webkit-box, -webkit-inline-box]
